requirements.txt
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
PyJWT==2.9.0
python-decouple==3.8

